Development/TSH/database/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
flask-cors==4.0.0
